Подчистит Header
